+++ common/src/main/java/me/drex/antixray/util/Util.java
@@ -1,6 +1,7 @@
 package me.drex.antixray.util;
 
 import me.drex.antixray.interfaces.ILevel;
+import me.drex.antixray.util.controller.ChunkPacketBlockController;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelHeightAccessor;
 import net.minecraft.world.level.chunk.ChunkAccess;