+++ common/src/main/java/me/drex/antixray/util/controller/ObfuscateLayerChunkPacketBlockController.java
@@ -1,0 +1,23 @@
+package me.drex.antixray.util.controller;
+
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+
+import java.util.Set;
+import java.util.concurrent.Executor;
+import java.util.function.IntSupplier;
+
+public class ObfuscateLayerChunkPacketBlockController extends ObfuscateChunkPacketBlockController {
+
+    public ObfuscateLayerChunkPacketBlockController(Level level, Executor executor, Set<Block> replacementBlocks, Set<Block> hiddenBlocks, int maxBlockHeight, int updateRadius, boolean lavaObscures) {
+        super(level, executor, replacementBlocks, hiddenBlocks, maxBlockHeight, updateRadius, lavaObscures);
+    }
+
+    @Override
+    public IntSupplier layerIntSupplier(int numberOfBlocks) {
+        // Get ONE random int per layer
+        int result = super.layerIntSupplier(numberOfBlocks).getAsInt();
+        return () -> result;
+    }
+
+}