+++ CHANGELOG.md
@@ -5,11 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Block tag support (in configs)
+- Layer obfuscation (Engine Mode 3)
+
 ### Removed
 - Bypass permission: "antixray.bypass"
 
 ### Changed
 - Relocate forge dependencies
+- Improved default config
 
 ## [1.2.8] - 2022-12-19
 ### Fixed